app对接

+++ user/src/main/java/com/wanghuiwen/user/dao/NotificationGradeMapper.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface NotificationGradeMapper extends ApiMapper<NotificationGrade> {
-    void deleteByNotification(Integer notificationId);
+    void deleteByNotification(Long notificationId);
 }
